My Development
I am going to upload every file that I worked on while i was taking this
class. Since the project was cumulative, I should be able to generate
the same CLASS files that are in the Compiler directory using my
uploaded parts.

This first commit includes the root directory for upcoming commits as
well as my Token Grammar for the compiler.