My Grammar
I am going to upload every file that I worked on while i was taking this
class. Since the project was cumulative, I should be able to generate
the same CLASS files that are in the Compiler directory using my
uploaded parts.

This second commit includes the Grammar that I developed for this
compiler. The job of the Grammar is to act as a parser for the entire
language.