Beakar completed parts
I am going to upload every file that I worked on while i was taking this
class. Since the project was cumulative, I should be able to generate
the same CLASS files that are in the Compiler directory using my
uploaded parts.

This commit contains the files that I worked on for Code Generation,
Name Binding, and Type checking in their respective subdirectories